email_senders/daily_email/requirements.txt
mysql-connector==2.1.6
python-http-client==3.1.0
sendgrid==5.6.0
